+++ packages/markets/src/components/expandMarkets/expandMarkets.ui.tsx
@@ -1,10 +1,16 @@
 import React from "react";
-import { useTranslation } from "@orderly.network/i18n";
 import { Box, cn, TabPanel, Tabs } from "@orderly.network/ui";
-import { FavoritesIcon } from "../../icons";
 import { MarketsTabName } from "../../type";
+import { useMarketsContext } from "../marketsProvider";
 import { RwaTab } from "../rwaTab";
 import { useFavoritesProps } from "../shared/hooks/useFavoritesExtraProps";
+import {
+  isBuiltInMarketTab,
+  tabKey,
+  resolveTabTitle,
+  useBuiltInTitles,
+  useCustomTabDataFilters,
+} from "../shared/tabUtils";
 import type { ExpandMarketsScriptReturn } from "./expandMarkets.script";
 
 const LazySearchInput = React.lazy(() =>
@@ -26,32 +32,53 @@ const cls = "oui-h-[calc(100%_-_36px)]";
 export const ExpandMarkets: React.FC<ExpandMarketsProps> = (props) => {
   const { activeTab, onTabChange, tabSort, onTabSort } = props;
 
-  const { t } = useTranslation();
-
   const { getFavoritesProps, renderEmptyView } = useFavoritesProps();
+  const builtInTitles = useBuiltInTitles();
+  const { tabs } = useMarketsContext();
+  const tabDataFilters = useCustomTabDataFilters(tabs);
 
-  const renderTab = (type: MarketsTabName) => {
+  const renderBuiltInContent = (type: string) => {
+    const tabType = type as MarketsTabName;
     return (
       <div className={cls}>
         <React.Suspense fallback={null}>
           <LazyMarketsListWidget
-            type={type}
+            type={tabType}
             initialSort={tabSort[type]}
-            onSort={onTabSort(type)}
+            onSort={onTabSort(tabType)}
             tableClassNames={{
               root: "oui-expandMarkets-list",
               scroll: cn(
                 "oui-px-1",
-                type === MarketsTabName.Favorites ? "oui-pb-9" : "oui-pb-2",
+                tabType === MarketsTabName.Favorites ? "oui-pb-9" : "oui-pb-2",
               ),
             }}
-            {...getFavoritesProps(type)}
+            {...getFavoritesProps(tabType)}
             emptyView={renderEmptyView({
-              type,
+              type: tabType,
               onClick: () => {
                 onTabChange(MarketsTabName.All);
               },
             })}
+          />
+        </React.Suspense>
+      </div>
+    );
+  };
+
+  const renderCustomContent = (key: string) => {
+    return (
+      <div className={cls}>
+        <React.Suspense fallback={null}>
+          <LazyMarketsListWidget
+            type={MarketsTabName.All}
+            dataFilter={(data) => tabDataFilters[key]?.(data) ?? data}
+            initialSort={tabSort[key]}
+            onSort={onTabSort(key as MarketsTabName)}
+            tableClassNames={{
+              root: "oui-expandMarkets-list",
+              scroll: cn("oui-px-1", "oui-pb-2"),
+            }}
           />
         </React.Suspense>
       </div>
@@ -86,56 +113,24 @@ export const ExpandMarkets: React.FC<ExpandMarketsProps> = (props) => {
         className={cn("oui-expandMarkets-tabs", cls)}
         showScrollIndicator
       >
-        <TabPanel
-          classNames={{
-            trigger: "oui-tabs-favorites-trigger",
-            content: "oui-tabs-favorites-content",
-          }}
-          title={<FavoritesIcon />}
-          value={MarketsTabName.Favorites}
-        >
-          {renderTab(MarketsTabName.Favorites)}
-        </TabPanel>
-        <TabPanel
-          classNames={{
-            trigger: "oui-tabs-all-trigger",
-            content: "oui-tabs-all-content",
-          }}
-          title={t("common.all")}
-          value={MarketsTabName.All}
-        >
-          {renderTab(MarketsTabName.All)}
-        </TabPanel>
-        <TabPanel
-          classNames={{
-            trigger: "oui-tabs-rwa-trigger",
-            content: "oui-tabs-rwa-content",
-          }}
-          title={<RwaTab />}
-          value={MarketsTabName.Rwa}
-        >
-          {renderTab(MarketsTabName.Rwa)}
-        </TabPanel>
-        <TabPanel
-          classNames={{
-            trigger: "oui-tabs-newListings-trigger",
-            content: "oui-tabs-newListings-content",
-          }}
-          title={t("markets.newListings")}
-          value={MarketsTabName.NewListing}
-        >
-          {renderTab(MarketsTabName.NewListing)}
-        </TabPanel>
-        <TabPanel
-          classNames={{
-            trigger: "oui-tabs-recent-trigger",
-            content: "oui-tabs-recent-content",
-          }}
-          title={t("markets.recent")}
-          value={MarketsTabName.Recent}
-        >
-          {renderTab(MarketsTabName.Recent)}
-        </TabPanel>
+        {tabs?.map((tab, index) => {
+          const key = tabKey(tab, index);
+          return (
+            <TabPanel
+              key={key}
+              classNames={{
+                trigger: `oui-tabs-${key}-trigger`,
+                content: `oui-tabs-${key}-content`,
+              }}
+              title={resolveTabTitle(tab, builtInTitles, <RwaTab />)}
+              value={key}
+            >
+              {isBuiltInMarketTab(tab)
+                ? renderBuiltInContent(tab.type)
+                : renderCustomContent(key)}
+            </TabPanel>
+          );
+        })}
       </Tabs>
     </Box>
   );