+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,18 @@
 # tsconfig
 
+## 0.14.2
+
+### Patch Changes
+
+- publish
+- 1c40427: publish
+
+## 0.14.2-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.14.1
 
 ### Patch Changes