+++ packages/ui-share/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/ui-share
 
+## 2.11.2
+
+### Patch Changes
+
+- publish
+- 1c40427: publish
+- Updated dependencies
+- Updated dependencies [1c40427]
+  - @orderly.network/hooks@2.11.2
+  - @orderly.network/i18n@2.11.2
+  - @orderly.network/types@2.11.2
+  - @orderly.network/ui@2.11.2
+  - @orderly.network/utils@2.11.2
+
+## 2.11.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/hooks@2.11.2-alpha.0
+  - @orderly.network/i18n@2.11.2-alpha.0
+  - @orderly.network/types@2.11.2-alpha.0
+  - @orderly.network/ui@2.11.2-alpha.0
+  - @orderly.network/utils@2.11.2-alpha.0
+
 ## 2.11.1
 
 ### Patch Changes